move poi test to models
