+++ app/src/androidTest/java/com/github/epfl/meili/forum/ForumActivityTest.kt
@@ -24,8 +24,6 @@ import com.github.epfl.meili.map.MapActivity
 import com.github.epfl.meili.models.PointOfInterest
 import com.github.epfl.meili.models.Post
 import com.github.epfl.meili.photo.CameraActivity
-import com.github.epfl.meili.review.ReviewsActivity
-import com.github.epfl.meili.review.ReviewsActivityTest
 import com.github.epfl.meili.storage.FirebaseStorageService
 import com.google.android.gms.tasks.Task
 import com.google.firebase.firestore.*
