+++ app/src/androidTest/java/com/github/epfl/meili/map/PoiMarkerViewModelTest.kt
@@ -4,8 +4,8 @@ import android.location.Location
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.internal.runner.junit4.statement.UiThreadStatement
 import com.github.epfl.meili.database.Database
-import com.github.epfl.meili.poi.PoiService
 import com.github.epfl.meili.models.PointOfInterest
+import com.github.epfl.meili.poi.PoiService
 import com.google.android.gms.maps.model.LatLng
 import com.google.firebase.firestore.*
 import org.junit.Assert.assertEquals
