+++ app/src/main/java/com/github/epfl/meili/map/MapActivity.kt
@@ -14,8 +14,8 @@ import com.github.epfl.meili.R
 import com.github.epfl.meili.database.FirestoreDatabase
 import com.github.epfl.meili.forum.ForumActivity
 import com.github.epfl.meili.home.Auth
-import com.github.epfl.meili.poi.PoiService
 import com.github.epfl.meili.models.PointOfInterest
+import com.github.epfl.meili.poi.PoiService
 import com.google.android.gms.location.FusedLocationProviderClient
 import com.google.android.gms.location.LocationServices
 import com.google.android.gms.maps.CameraUpdateFactory.newLatLngZoom
