+++ app/src/main/java/com/github/epfl/meili/messages/ChatLogActivity.kt
@@ -16,8 +16,8 @@ import com.github.epfl.meili.forum.ForumActivity
 import com.github.epfl.meili.home.Auth
 import com.github.epfl.meili.map.MapActivity
 import com.github.epfl.meili.models.ChatMessage
-import com.github.epfl.meili.models.User
 import com.github.epfl.meili.models.PointOfInterest
+import com.github.epfl.meili.models.User
 import com.github.epfl.meili.review.ReviewsActivity
 import com.github.epfl.meili.util.DateAuxiliary
 import com.xwray.groupie.GroupAdapter
