+++ app/src/test/java/com/github/epfl/meili/models/PointOfInterestTest.kt
@@ -1,6 +1,5 @@
-package com.github.epfl.meili.poi
+package com.github.epfl.meili.models
 
-import com.github.epfl.meili.models.PointOfInterest
 import com.google.android.gms.maps.model.LatLng
 import org.junit.Assert.assertEquals
 import org.junit.Test
